requirements.txt
fastapi
uvicorn[standard]
pandas
matplotlib
seaborn
scikit-learn
python-multipart
langchain
langchain-google-genai
statsmodels
scipy
joblib
PyMuPDF
faiss-cpu
langchain-community
tinydb
python-dotenv
langchain-core
pydantic
langchain-groq
langchain-openai